refactor: extract common renderer functionality
